requirements-agent.txt
playwright==1.49.0
tf-playwright-stealth==1.1.2
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
